Update payment styling (#9)
* use library wallet model

* use lib transfer for transaction

* add amount input formatter

* use mock transaction

* use entities wallet model instead of lib wallet model

* use payment screen and irouter

* adapt new user flow. use payment screen

* update font from trial to pro

* udpate scan screen cancel label

* add isShop flag to main layout

* update rhombus icon button draft

* update payment screen styles

* style success screen. rename view_state folder

* fix injection problem. rename Wallet to WalletEntity

+++ lib/ui/core/services/abstract_qr_scanner.dart
@@ -1,11 +1,12 @@
 import 'package:dartz/dartz.dart';
 import 'package:e_coupon/business/core/failure.dart';
+import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
 // maybe extend transaction?
 class ScannedResult {
   final String walletID;
-  final double amount;
+  final int amount;
 
   ScannedResult({@required this.walletID, this.amount});
 }
